[#1] Add special case for incorrectly formatted .SRT files
OPTIONS3 and TITLES3 seem to use an earlier version of the file format, and so parses incorrectly.
Even vanilla Sonic R chokes on these when attempting to finish a race on these "tracks".
I've set a reasonable maximum number of replay cam/Sec9 points to try and work around this.